Updat eon 0-print_list.c

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t j = 0;
+	int j = 0;
 
 	while (h != NULL)
 	{
